+++ src/cogs/numbers.py
@@ -5,7 +5,7 @@ from utils.bot import Bot
 import nextcord
 
 from utils.data import get_server_path
-from utils.database import get_channel
+from utils.database import get_channel_of_type
 
 JSON_CONTENT = {"number": 0, "last_user": None, "record": 0}
 NUMBERS_JSON = "numbers.json"
@@ -68,7 +68,7 @@ class numbers(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message: nextcord.Message):
-        if get_channel(message.guild, "numbers") != message.channel.id:
+        if get_channel_of_type(message.guild, "numbers") != message.channel.id:
             return
         content = message.content
         number = self.get_number(message.guild)