+++ Changelog.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Añadido
 - Al usar los comandos addpost y setup se comprobará si el bot tiene acceso a los canales que se especifiquen y si tiene permiso para enviar mensajes.
 - Comprobacion de permisos de canal en task post
+- Comando setchannelpolicy. Con este comando se puede establecer que tipo de contenido se puede enviar en un canal. Si un mensaje no tiene ese tipo de contenido se eliminará.
+- Comando setchannel. Establecer un canal para que sea un canal predefinido. Por ejemplo en el canal lobby se da la bienvenida a los miembros nuevos. Los canales predefinidos que hay son lobby, general, audit, noticias, games, wordle, numbers y ordure
+- Comando setchannels. Parecido a setchannel pero permite ir estableciendo los canales uno a uno.
+- Eliminación de contenido que no cumpla las politicas de un canal 
 
 ### Arreglado
 - Task joined_date que felicita a los usuarios al cumplir x años en el servidor